all are working

+++ src/components/Card.jsx
@@ -4,13 +4,18 @@ import { BsThreeDots } from "react-icons/bs";
 import { IoEyeOutline, IoEyeOffOutline } from "react-icons/io5";
 import { MdOutlineModeEdit, MdDeleteOutline } from "react-icons/md"
 import { useBlog } from '../context/BlogContext';
+import { useNavigate } from 'react-router';
+import { toast } from 'react-toastify';
 
 const Card = ({ blog }) => {
-    const { deleteBlog, togglePublish } = useBlog();
+    const { deleteBlog, togglePublish, setEditBlog } = useBlog();
     const { theme } = useTheme();
     const [menu, setMenu] = useState(false);
+    const navigate = useNavigate();
 
     const handleDelete = (id) => {
+        const confirmDelete = window.confirm("Are you sure you want to delete this blog?");
+        if (!confirmDelete) return;
         deleteBlog(id);
         setMenu(p => !p);
     }
@@ -19,6 +24,13 @@ const Card = ({ blog }) => {
         togglePublish(id);
         setMenu(p => !p);
     }
+
+    const handleEdit = () => {
+        setEditBlog(blog);
+        setMenu(p => !p);
+        navigate('/dashboard/new');
+
+    };
 
     return (
         < div className='mt-5 border border-zinc-800 rounded-xl p-6 flex justify-between items-center' >
@@ -42,7 +54,7 @@ const Card = ({ blog }) => {
                 <button onClick={() => setMenu(prev => !prev)} className='text-xl p-2 rounded-lg hover:bg-blue-600 cursor-pointer duration-200'><BsThreeDots size={18} /></button>
                 {/* Menu */}
                 <div className={` ${menu ? 'block' : 'hidden'} text-left p-1 rounded-lg border z-5 ${theme ? 'bg-zinc-900' : 'bg-zinc-200'} top-12 right-0 absolute`}>
-                    <div className='hover:bg-blue-600 rounded-lg cursor-pointer duration-200 px-4 py-2'>
+                    <div onClick={() => handleEdit(blog.id)} className='hover:bg-blue-600 rounded-lg cursor-pointer duration-200 px-4 py-2'>
                         <h1 className='flex items-center gap-2'><MdOutlineModeEdit size={15} /> Edit</h1>
                     </div>
                     <div onClick={() => handlePublish(blog.id)} className='hover:bg-blue-600 rounded-lg cursor-pointer duration-200 px-4 py-2'>
